backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=43.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
pydantic>=2.10.0
pydantic-settings>=2.1.0
easyocr>=1.7.1
rapidocr-onnxruntime>=1.3.22
opencv-python-headless>=4.10.0.84
Pillow>=11.0.0
gTTS>=2.5.0
aiofiles>=23.2.1
apscheduler>=3.10.4
numpy>=1.26.2
